+++ terezinha-farm/api/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.3](https://github.com/ttoss/ttoss/compare/@terezinha-farm/api@0.6.2...@terezinha-farm/api@0.6.3) (2023-01-19)
+
+**Note:** Version bump only for package @terezinha-farm/api
+
 ## [0.6.2](https://github.com/ttoss/ttoss/compare/@terezinha-farm/api@0.6.1...@terezinha-farm/api@0.6.2) (2023-01-14)
 
 **Note:** Version bump only for package @terezinha-farm/api